Change typedefs of ptrdiff_t, size_t and wchar_t. Add ssize_t.

+++ trunk/dingoo_sdk/src/libc/stddef.h
@@ -1,12 +1,14 @@
 #ifndef __stddef_h__
 #define __stddef_h__
 
+#include <stdint.h>
+
 #ifndef NULL
 #define NULL ((void *)0)
 #endif
 
-typedef long int ptrdiff_t;
-typedef unsigned long int size_t;
-typedef unsigned short wchar_t;
+typedef intptr_t ptrdiff_t;
+typedef uintptr_t size_t;
+typedef uint16_t wchar_t;
 
 #endif
